Moved reusable functions to utils
