+++ src/components/tab-panel/TabPanel.tsx
@@ -1,0 +1,31 @@
+import { Box, Typography } from "@mui/material";
+
+interface TabPanelProps {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+  style?: any;
+}
+
+const TabPanel = (props: TabPanelProps) => {
+  const { children, value, index, style = {}, ...other } = props;
+
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      style={style}
+      {...other}
+    >
+      {value === index && (
+        <Box>
+          <Typography component={"span"}>{children}</Typography>
+        </Box>
+      )}
+    </div>
+  );
+};
+
+export default TabPanel;
