+++ src/containers/groups/components/create-edit/CreateEditGroup.tsx
@@ -3,8 +3,9 @@ import { useRecoilState } from "recoil";
 import { useNavigate, useParams } from "react-router-dom";
 import { FieldValues } from "react-hook-form";
 import { useSetRecoilState } from "recoil";
-import { Box, Tab, Tabs, Typography, Grid, Divider } from "@mui/material";
+import { Box, Tab, Tabs, Grid, Divider } from "@mui/material";
 import CircularProgress from "@mui/material/CircularProgress";
+import { useLazyQuery } from "@apollo/client";
 
 import { GET_ROLES } from "containers/roles/services/queries";
 import {
@@ -39,34 +40,7 @@ import { CustomAvatar } from "components/custom-avatar/CustomAvatar";
 import { ReactComponent as CrossIcon } from "assets/cross-icon.svg";
 import { useCustomQuery } from "hooks/useQuery";
 import { useCustomMutation } from "hooks/useMutation";
-import { useLazyQuery } from "@apollo/client";
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-  style?: any;
-}
-
-function TabPanel(props: TabPanelProps) {
-  const { children, value, index, style = {}, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      style={style}
-      {...other}
-    >
-      {value === index && (
-        <Box>
-          <Typography component={"span"}>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  );
-}
+import TabPanel from "components/tab-panel/TabPanel";
 
 const CreateOrEditGroup = () => {
   const { id } = useParams();
@@ -175,7 +149,6 @@ const CreateOrEditGroup = () => {
         return;
       }
       if (item) {
-        // handlePermissions(item);
         if (roles[0] === null) {
           setRoles([item]);
         } else {
